Fix test_effective_decrease_balance_updates_lookahead reftest
